Add recordedField component
